[runtime][common][assembler][cmake] Refactor
- Refactor project structure
