+++ runtime/CMakeLists.txt
@@ -1,18 +1,17 @@
-set(RUNTIME_SRC
-	runtime.cpp
+# Define runtime library build
+
+add_library(runtime STATIC)
+add_library(shrimp::runtime ALIAS runtime)
+
+add_subdirectory(common)
+
+# Add runtime_runtime include and src
+add_subdirectory(runtime)
+# Add runtime_shrimp_vm include and src
+add_subdirectory(shrimp_vm)
+
+target_link_libraries(runtime
+PUBLIC
+    shrimp::common
+    shrimp::runtime_common
 )
-
-add_library(runtime STATIC ${RUNTIME_SRC})
-
-target_link_libraries(runtime PRIVATE interpreter)
-
-set(RUNTIME_DIR ${PROJECT_ROOT}/runtime)
-
-target_include_directories(runtime PUBLIC
-	${RUNTIME_DIR}
-	${RUNTIME_DIR}/interpreter
-	${PROJECT_ROOT}/common
-)
-
-
-add_subdirectory(interpreter)
