--- runtime/interpreter/CMakeLists.txt
@@ -1,7 +1,0 @@
-add_library(interpreter STATIC interpreter.cpp intrinsics.cpp)
-
-target_include_directories(interpreter PUBLIC
-	${RUNTIME_DIR}
-	${RUNTIME_DIR}/interpreter
-	${PROJECT_SOURCE_DIR}/common
-)
