+++ shrimp/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Defines shrimp VM build
+
+add_executable(shrimp src/shrimp.cpp)
+
+target_link_libraries(shrimp PRIVATE shrimp::runtime)
